♻️ [Refactoring]: files APIを関数ベースに変更
- file と file-nodes を独立したAPIとして分離
- 各APIを関数ベースの実装に変更（getFileApi, getFileNodesApi）
- 後方互換性のため files-compat.ts を追加
- テストファイルを __tests__ ディレクトリに整理
- ESLintエラーを修正（as any を削除）
