+++ .devcontainer/Dockerfile
@@ -36,8 +36,8 @@ RUN apt-get update && apt-get install -y \
     && update-locale LANG=ja_JP.UTF-8 \
     && rm -rf /var/lib/apt/lists/*
 
-# Install Node.js 20.x
-RUN curl -fsSL https://deb.nodesource.com/setup_20.x | bash - \
+# Install Node.js 22.x
+RUN curl -fsSL https://deb.nodesource.com/setup_22.x | bash - \
     && apt-get install -y nodejs \
     && rm -rf /var/lib/apt/lists/*
