+++ package.json
@@ -26,10 +26,12 @@
     "zod": "^3.25.76"
   },
   "devDependencies": {
+    "@mizchi/lsmcp": "^0.9.4",
     "@types/express": "5.0.3",
     "@types/node": "20.11.0",
     "@typescript-eslint/eslint-plugin": "6.18.1",
     "@typescript-eslint/parser": "6.18.1",
+    "@typescript/native-preview": "^7.0.0-dev.20250825.1",
     "eslint": "8.56.0",
     "eslint-config-prettier": "9.1.2",
     "express": "5.1.0",
