+++ src/api/figma-api-client.ts
@@ -1,6 +1,8 @@
 import { createApiConfig } from './config.js';
 import { createHttpClient, type HttpClient } from './client.js';
-import { createFilesApi, type FilesApi } from './endpoints/files/index.js';
+import { getFileApi } from './endpoints/file/index.js';
+import { getFileNodesApi } from './endpoints/file-nodes/index.js';
+import { createFilesApi, type FilesApi } from './endpoints/files-compat.js';
 import { createNodesApi } from './endpoints/nodes/index.js';
 import { fileComponentsApi, fileComponentSetsApi } from './endpoints/components/index.js';
 import { createStylesApi } from './endpoints/styles/index.js';
@@ -20,6 +22,7 @@ import type { ExportImageResponse } from '../types/api/responses/image-responses
 import type { GetFileCommentsApiResponse } from '../types/api/responses/comment-responses.js';
 import type { GetVersionsResponse } from '../types/api/responses/version-responses.js';
 import type { ExportImageOptions } from '../types/api/options/image-options.js';
+import type { FigmaFile, GetFileOptions, GetFileNodesResponse } from '../types/index.js';
 
 /**
  * FigmaApiClientのデータ構造
@@ -177,6 +180,31 @@ export namespace FigmaApiClient {
   ): Promise<GetVersionsResponse> {
     const response = await client.versions.getVersions(fileKey);
     return convertKeysToCamelCase(response);
+  }
+
+  /**
+   * ファイル情報を取得（キャメルケース変換付き）
+   */
+  export async function getFile(
+    client: FigmaApiClient,
+    fileKey: string,
+    options?: GetFileOptions
+  ): Promise<FigmaFile> {
+    const response = await getFileApi(client.httpClient, fileKey, options);
+    return convertKeysToCamelCase(response) as FigmaFile;
+  }
+
+  /**
+   * ファイルノード情報を取得（キャメルケース変換付き）
+   */
+  export async function getFileNodes(
+    client: FigmaApiClient,
+    fileKey: string,
+    ids: string[],
+    options?: GetFileOptions
+  ): Promise<GetFileNodesResponse> {
+    const response = await getFileNodesApi(client.httpClient, fileKey, ids, options);
+    return convertKeysToCamelCase(response) as GetFileNodesResponse;
   }
 }
